:sparkles: separate file for Request Filter
